Data.Functor.Contravariant.Representable

Description

Representable contravariant endofunctors over the category of Haskell types are isomorphic to (_ -> r) and resemble mappings to a fixed range.

Values

type family Value f

Contravariant Keyed

class Contravariant f => Valued f where

Dual to Keyed.

Methods

contramapWithValue :: (b -> Either a (Value f)) -> f a -> f b

Contravariant Indexed

class Coindexed f where

Dual to Indexed.

Methods

coindex :: f a -> a -> Value f

Representable Contravariant Functors

class (Coindexed f, Valued f) => Representable f where

A Contravariant functor f is Representable if contrarep and coindex witness an isomorphism to (_ -> Value f).

Methods

contrarep :: (a -> Value f) -> f a

 contramap f (contrarep g) = contrarep (g . f)

Default definitions

contramapDefault :: Representable f => (a -> b) -> f b -> f a

contramapWithValueDefault :: Representable f => (b -> Either a (Value f)) -> f a -> f b