Data.Functor.Representable

Description

Representable endofunctors over the category of Haskell types are isomorphic to the reader monad and so inherit a very large number of properties for free.

Representable Functors

class (Distributive f, Indexable f) => Representable f where

A Functor f is Representable if tabulate and index witness an isomorphism to (->) x.

 tabulate . index = id
 index . tabulate = id
 tabulate . return f = return f

Methods

tabulate :: (Key f -> a) -> f a

 fmap f . tabulate = tabulate . fmap f

Wrapped representable functors

newtype Rep f a

Constructors

Rep

Fields

unrep :: f a

Default definitions

Functor

fmapRep :: Representable f => (a -> b) -> f a -> f b

Distributive

distributeRep :: (Representable f, Functor w) => w (f a) -> f (w a)

Keyed

mapWithKeyRep :: Representable f => (Key f -> a -> b) -> f a -> f b

Apply/Applicative

apRep :: Representable f => f (a -> b) -> f a -> f b

pureRep :: Representable f => a -> f a

liftR2 :: Representable f => (a -> b -> c) -> f a -> f b -> f c

liftR3 :: Representable f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d

Bind/Monad

bindRep :: Representable f => f a -> (a -> f b) -> f b

bindWithKeyRep :: Representable f => f a -> (Key f -> a -> f b) -> f b

Zip/ZipWithKey

zipWithRep :: Representable f => (a -> b -> c) -> f a -> f b -> f c

zipWithKeyRep :: Representable f => (Key f -> a -> b -> c) -> f a -> f b -> f c

MonadReader

askRep :: Representable f => f (Key f)

localRep :: Representable f => (Key f -> Key f) -> f a -> f a

Extend

duplicatedRep :: (Representable f, Semigroup (Key f)) => f a -> f (f a)

extendedRep :: (Representable f, Semigroup (Key f)) => (f a -> b) -> f a -> f b

Comonad

duplicateRep :: (Representable f, Monoid (Key f)) => f a -> f (f a)

extendRep :: (Representable f, Monoid (Key f)) => (f a -> b) -> f a -> f b

extractRep :: (Indexable f, Monoid (Key f)) => f a -> a